test(e2e): split DPoP and RAR suites by RFC concern for finer parallelism
The class-level parallelism from the previous commit was bounded by the
two largest classes (DPoPTests ~4.4s, RichAuthorizationRequestsTests
~3.4s) since xunit serialises methods within a class. Split each along
its natural RFC seams into focused classes sharing a non-test base:

  DPoPTests         -> DPoPTests (§6 token endpoint + discovery),
                       DPoPRefreshTests (§5), DPoPUserInfoTests (§9)
                       over DPoPTestBase
  RichAuthorization -> RichAuthorizationRequestsTests (round-trip +
   RequestsTests        validation), RarMetadataTests (discovery/DCR/
                        license/introspection), RarConsentTests (§5
                        consent + refresh) over RarTestBase

Each runs as its own parallel collection. Suite wall-clock drops from
~5s to ~3s (all 48 tests green, stable across repeated runs).
